requirements.txt
numpy>=1.21.0
torch>=1.9.0
librosa>=0.9.0
transformers>=4.11.0
matplotlib>=3.4.0
umap-learn>=0.5.0
scikit-learn>=0.24.0

